+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+**Aug 2024**
+
+* Update baseline and nutrient data
+
 **Feb 2023**
 
 * Add buttons to download entire baseline and nutrient datasets